Rebased

+++ ocs_ci/ocs/dr/dr_workload.py
@@ -249,19 +249,9 @@ class BusyBox(DRWorkload):
                     resource_type="namespace", resource_description="busybox-workloads"
                 )
             )
+            self.workload_namespace = workload_ns_yaml_data["metadata"]["name"]
             templating.dump_data_to_temp_yaml(
                 workload_ns_yaml_data, self.namespace_yaml_file
-            )
-
-            # load placementrule
-            placementrule_yaml_data = templating.load_yaml(self.placementrule_yaml_file)
-            placementrule_yaml_data["metadata"]["name"] = (
-                helpers.create_unique_resource_name(
-                    resource_type="placementrule", resource_description="busybox"
-                )
-            )
-            templating.dump_data_to_temp_yaml(
-                placementrule_yaml_data, self.placementrule_yaml_file
             )
 
             # load drpc.yaml
@@ -269,12 +259,68 @@ class BusyBox(DRWorkload):
             drpc_yaml_data["metadata"]["name"] = helpers.create_unique_resource_name(
                 resource_type="drpc", resource_description="busybox"
             )
-            drpc_yaml_data["spec"]["placementRef"]["name"] = placementrule_yaml_data[
-                "metadata"
-            ]["name"]
             drpc_yaml_data["spec"]["preferredCluster"] = cluster
             drpc_yaml_data["spec"]["drPolicyRef"]["name"] = self.dr_policy_name
             templating.dump_data_to_temp_yaml(drpc_yaml_data, self.drpc_yaml_file)
+
+            if self.is_placement:
+                # load placement.yaml
+                clusterset_name = (
+                    config.ENV_DATA.get("cluster_set") or get_cluster_set_name()[0]
+                )
+                placement_yaml_data = templating.load_yaml(self.placement_yaml_file)
+                placement_yaml_data["spec"]["predicates"][0]["requiredClusterSelector"][
+                    "labelSelector"
+                ]["matchExpressions"][0]["values"][0] = self.preferred_primary_cluster
+                placement_yaml_data["spec"]["clusterSets"][0] = clusterset_name
+                self.sub_placement_name = placement_yaml_data["metadata"]["name"]
+                templating.dump_data_to_temp_yaml(
+                    placement_yaml_data, self.placement_yaml_file
+                )
+                managed_clusterset_binding_yaml_data = templating.load_yaml(
+                    self.managed_clusterset_binding_file
+                )
+                managed_clusterset_binding_yaml_data["metadata"][
+                    "namespace"
+                ] = self.workload_namespace
+                managed_clusterset_binding_yaml_data["metadata"][
+                    "name"
+                ] = clusterset_name
+                managed_clusterset_binding_yaml_data["spec"][
+                    "clusterSet"
+                ] = clusterset_name
+                templating.dump_data_to_temp_yaml(
+                    managed_clusterset_binding_yaml_data,
+                    self.managed_clusterset_binding_file,
+                )
+                if placement_yaml_data["kind"] == "Placement":
+                    drpc_yaml_data = templating.load_yaml(self.drpc_yaml_file_placement)
+                    drpc_yaml_data["metadata"][
+                        "name"
+                    ] = f"{self.sub_placement_name}-drpc"
+                    drpc_yaml_data["spec"][
+                        "preferredCluster"
+                    ] = self.preferred_primary_cluster
+                    drpc_yaml_data["spec"]["drPolicyRef"]["name"] = self.dr_policy_name
+                    drpc_yaml_data["spec"]["placementRef"][
+                        "name"
+                    ] = self.sub_placement_name
+
+                    drpc_yaml_data["metadata"]["namespace"] = self.workload_namespace
+                    drpc_yaml_data["spec"]["placementRef"][
+                        "namespace"
+                    ] = self.workload_namespace
+                    drpc_yaml_data["spec"]["pvcSelector"][
+                        "matchLabels"
+                    ] = self.workload_pvc_selector
+                    del drpc_yaml_data["spec"]["pvcSelector"]["matchExpressions"]
+                    del drpc_yaml_data["spec"]["kubeObjectProtection"]
+                    self.drpc_data_yaml = tempfile.NamedTemporaryFile(
+                        mode="w+", prefix="drpc", delete=False
+                    )
+                    templating.dump_data_to_temp_yaml(
+                        drpc_yaml_data, self.drpc_data_yaml.name
+                    )
 
             # load channel.yaml
             channel_yaml_data = templating.load_yaml(self.channel_yaml_file)
@@ -306,7 +352,7 @@ class BusyBox(DRWorkload):
                 + channel_yaml_data["metadata"]["name"]
             )
             subscription_yaml_data["spec"]["placement"]["placementRef"]["name"] = (
-                placementrule_yaml_data["metadata"]["name"]
+                placement_yaml_data["metadata"]["name"]
             )
             templating.dump_data_to_temp_yaml(
                 subscription_yaml_data, self.subscription_yaml_file
@@ -349,7 +395,9 @@ class BusyBox(DRWorkload):
             run_cmd(
                 f"oc create -k {self.workload_subscription_dir}/{self.workload_name}"
             )
-
+            if self.is_placement:
+                self.add_annotation_to_placement()
+                run_cmd(f"oc create -f {self.drpc_data_yaml.name}")
             self.verify_workload_deployment(cluster)
 
     def _deploy_prereqs(self):
@@ -1199,18 +1247,25 @@ class BusyboxDiscoveredApps(DRWorkload):
         run_cmd(f"oc create -k {self.workload_path} -n {self.workload_namespace} ")
         self.check_pod_pvc_status(skip_replication_resources=True)
         config.switch_acm_ctx()
-        self.create_placement()
+        if not self.discovered_apps_multi_ns:
+            self.create_placement()
+
         if recipe:
             log.info("Creating workload with recipe")
+
+            # Switch back to primary, then to each managed cluster to apply recipe
             config.switch_to_cluster_by_name(self.preferred_primary_cluster)
             for cluster in get_non_acm_cluster_config():
                 config.switch_ctx(cluster.MULTICLUSTER["multicluster_index"])
                 self.create_recipe_with_checkhooks()
+
             config.switch_acm_ctx()
             self.create_drpc_for_apps_with_recipe()
-        else:
+            self.verify_workload_deployment()
+
+        elif not self.discovered_apps_multi_ns:
             self.create_drpc()
-        self.verify_workload_deployment()
+            self.verify_workload_deployment()
 
     def _deploy_prereqs(self):
         """
@@ -1318,7 +1373,7 @@ class BusyboxDiscoveredApps(DRWorkload):
         log.info(f"Creating Placement for workload {self.workload_name}")
         run_cmd(f"oc create -f {placement_yaml.name}")
 
-    def create_dprc(
+    def create_drpc(
         self,
         drpc_name=None,
         placement_name=None,
@@ -1592,8 +1647,14 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
             "discovered_apps_pod_selector_value"
         )
 
-    def deploy_workload(self):
+    def deploy_workload(self, shared_drpc_protection=False, dr_protect=True):
         """
+
+        Args:
+            shared_drpc_protection (bool): False by default, another workload in an existing namespace
+                            will share the DRPC and DR protect it using Shared Protection type via UI
+            dr_protect (bool): True by default where workload will be DR protected via CLI,
+                                else test case should handle it (via UI)
 
         Deployment specific to CNV workload for Discovered/Imperative Apps
 
@@ -1601,16 +1662,23 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
         self._deploy_prereqs()
         for cluster in get_non_acm_cluster_config():
             config.switch_ctx(cluster.MULTICLUSTER["multicluster_index"])
-            self.create_namespace()
-        self.manage_dr_vm_secrets()
+            if not shared_drpc_protection:
+                self.create_namespace()
+            else:
+                # Shared=False means namespace exists, so skip creation
+                log.info(
+                    f"Namespace in use: {self.workload_namespace} for Shared Protection type"
+                )
+        self.manage_dr_vm_secrets(shared_drpc_protection=shared_drpc_protection)
         config.switch_to_cluster_by_name(self.preferred_primary_cluster)
         self.workload_path = self.target_clone_dir + "/" + self.workload_dir
         run_cmd(f"oc create -k {self.workload_path} -n {self.workload_namespace} ")
         self.check_pod_pvc_status(skip_replication_resources=True)
         config.switch_acm_ctx()
-        self.create_placement()
-        self.create_dprc()
-        self.verify_workload_deployment()
+        if dr_protect:
+            self.create_placement()
+            self.create_drpc()
+            self.verify_workload_deployment()
         self.vm_obj = VirtualMachine(
             vm_name=self.vm_name, namespace=self.workload_namespace
         )
@@ -1634,20 +1702,25 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
 
         run_cmd(f"oc create namespace {self.workload_namespace}")
 
-    def manage_dr_vm_secrets(self):
+    def manage_dr_vm_secrets(self, shared_drpc_protection=False):
         """
         Create secrets to access the VMs via SSH. If a secret already exists, delete and recreate it.
+
+        Args:
+            shared_drpc_protection (bool): False by default, True when Shared Protection type is used to DR Protect
+            a workload in an existing namespace
 
         """
         for cluster in get_non_acm_cluster_config():
             config.switch_ctx(cluster.MULTICLUSTER["multicluster_index"])
 
             # Create namespace if it doesn't exist
-            try:
-                create_project(project_name=self.workload_namespace)
-            except CommandFailed as ex:
-                if "(AlreadyExists)" in str(ex):
-                    log.warning("The namespace already exists!")
+            if not shared_drpc_protection:
+                try:
+                    create_project(project_name=self.workload_namespace)
+                except CommandFailed as ex:
+                    if "(AlreadyExists)" in str(ex):
+                        log.warning("The namespace already exists!")
 
             # Create or recreate the secret for ssh access
             try:
@@ -1699,7 +1772,7 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
         log.info(f"Creating Placement for workload {self.workload_name}")
         run_cmd(f"oc create -f {placement_yaml.name}")
 
-    def create_dprc(self):
+    def create_drpc(self):
         """
         Create DRPC for discovered Apps
 
@@ -1787,21 +1860,39 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
             skip_replication_resources=skip_replication_resources,
         )
 
-    def delete_workload(self):
+    def delete_workload(
+        self, shared_drpc_protection=False, skip_resource_deletion_verification=False
+    ):
         """
         Delete Discovered Apps
 
+        Args:
+            shared_drpc_protection (bool): False by default, another workload in an existing namespace
+                            will share the DRPC and DR protect it using Shared Protection type via UI
+            skip_resource_deletion_verification (bool): False by default, resource verification is
+                                                    handled based upon Standalone or Shared protection type is used
+                                                    for DR protection (via ACM UI)
+
         """
 
-        log.info("Deleting DRPC")
         config.switch_acm_ctx()
-        run_cmd(
-            f"oc delete drpc -n {constants.DR_OPS_NAMESAPCE} {self.discovered_apps_placement_name}"
-        )
-        log.info("Deleting Placement")
-        run_cmd(
-            f"oc delete placement -n {constants.DR_OPS_NAMESAPCE} {self.discovered_apps_placement_name}-placement-1"
-        )
+        if not shared_drpc_protection:
+            log.info("Deleting DRPC")
+            try:
+                run_cmd(
+                    f"oc delete drpc -n {constants.DR_OPS_NAMESAPCE} {self.discovered_apps_placement_name}"
+                )
+            except CommandFailed:
+                # This is needed when DRPolicy is applied via UI, where DRPC which is created has suffix -drpc
+                # hence deletion fails
+                run_cmd(
+                    f"oc delete drpc -n {constants.DR_OPS_NAMESAPCE} {self.discovered_apps_placement_name}-drpc"
+                )
+                log.info("DRPC deleted")
+            log.info("Deleting Placement")
+            run_cmd(
+                f"oc delete placement -n {constants.DR_OPS_NAMESAPCE} {self.discovered_apps_placement_name}-placement-1"
+            )
 
         for cluster in get_non_acm_cluster_config():
             config.switch_ctx(cluster.MULTICLUSTER["multicluster_index"])
@@ -1810,9 +1901,11 @@ class CnvWorkloadDiscoveredApps(DRWorkload):
                 f"oc delete -k {self.workload_path} -n {self.workload_namespace}",
                 ignore_error=True,
             )
-            dr_helpers.wait_for_all_resources_deletion(
-                namespace=self.workload_namespace,
-                discovered_apps=True,
-                workload_cleanup=True,
-            )
-            run_cmd(f"oc delete project {self.workload_namespace}")
+            if not skip_resource_deletion_verification:
+                dr_helpers.wait_for_all_resources_deletion(
+                    namespace=self.workload_namespace,
+                    discovered_apps=True,
+                    workload_cleanup=True,
+                )
+                run_cmd(f"oc delete project {self.workload_namespace}")
+                log.info(f"Project {self.workload_namespace} deleted successfully")
